+++ entrypoint.sh
@@ -38,10 +38,16 @@ unset SSMTP_HOST
 unset SSMTP_USER
 unset SSMTP_PASSWORD
 
+# Init
+echo "Initializing files and folders..."
+mkdir -p /data/db /data/action.d /data/filter.d /data/jail.d
+ln -sf /data/jail.d /etc/fail2ban/
+
 # Fail2ban conf
 echo "Setting Fail2ban configuration..."
 sed -i "s/logtarget =.*/logtarget = STDOUT/g" /etc/fail2ban/fail2ban.conf
 sed -i "s/loglevel =.*/loglevel = $F2B_LOG_LEVEL/g" /etc/fail2ban/fail2ban.conf
+sed -i "s/dbfile =.*/dbfile = \/data\/db\/fail2ban\.sqlite3/g" /etc/fail2ban/fail2ban.conf
 sed -i "s/dbpurgeage =.*/dbpurgeage = $F2B_DB_PURGE_AGE/g" /etc/fail2ban/fail2ban.conf
 cat > /etc/fail2ban/jail.local <<EOL
 [DEFAULT]
@@ -50,5 +56,29 @@ destemail = ${F2B_DEST_EMAIL}
 sender = ${F2B_SENDER}
 action = ${F2B_ACTION}
 EOL
+
+# Check custom actions
+echo "Checking for custom actions in /data/action.d..."
+actions=$(ls -l /data/action.d | egrep '^-' | awk '{print $9}')
+for action in ${actions}; do
+  if [ -f "/etc/fail2ban/action.d/${action}" ]; then
+    echo "  WARNING: ${action} already exists and will be overriden"
+    rm -f "/etc/fail2ban/action.d/${action}"
+  fi
+  echo "  Add custom action ${action}..."
+  ln -sf "/data/action.d/${action}" "/etc/fail2ban/action.d/"
+done
+
+# Check custom filters
+echo "Checking for custom filters in /data/filter.d..."
+filters=$(ls -l /data/filter.d | egrep '^-' | awk '{print $9}')
+for filter in ${filters}; do
+  if [ -f "/etc/fail2ban/filter.d/${filter}" ]; then
+    echo "  WARNING: ${filter} already exists and will be overriden"
+    rm -f "/etc/fail2ban/filter.d/${filter}"
+  fi
+  echo "  Add custom filter ${filter}..."
+  ln -sf "/data/filter.d/${filter}" "/etc/fail2ban/filter.d/"
+done
 
 exec "$@"